Added ability to exclude tile textures from the LRUCache
If textures are shared between tiles then currently the space that
the they take up is counted every time the texture is referenced.
Therefore this change allows a flag to be set on the tiles to exclude
their texture from the cache size calculation.

This does mean that the cache usage will now be underestimated but
we can track the texture usage separately, perhaps as part of a
virtual texturing system.